+++ apps/explorer/vite.config.ts
@@ -4,6 +4,14 @@ export default defineConfig({
   base: './',
   build: {
     target: 'es2022',
-    sourcemap: true
+    sourcemap: false,
+    modulePreload: { polyfill: false },
+    rollupOptions: {
+      output: {
+        entryFileNames: 'app.js',
+        chunkFileNames: '[name].js',
+        assetFileNames: '[name][extname]'
+      }
+    }
   }
 })